+++ Tests/Constraints/FileTest.php
@@ -14,6 +14,7 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Validator\Constraints\File;
 use Symfony\Component\Validator\Exception\ConstraintDefinitionException;
+use Symfony\Component\Validator\Exception\InvalidArgumentException;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
 use Symfony\Component\Validator\Mapping\Loader\AttributeLoader;
 
@@ -77,6 +78,31 @@ class FileTest extends TestCase
         }
 
         $this->assertSame(1000, $file->maxSize);
+    }
+
+    public function testFilenameMaxLength()
+    {
+        $file = new File(filenameMaxLength: 30);
+        $this->assertSame(30, $file->filenameMaxLength);
+    }
+
+    public function testDefaultFilenameCountUnitIsUsed()
+    {
+        $file = new File();
+        self::assertSame(File::FILENAME_COUNT_BYTES, $file->filenameCountUnit);
+    }
+
+    public function testFilenameCharsetDefaultsToNull()
+    {
+        $file = new File();
+        self::assertNull($file->filenameCharset);
+    }
+
+    public function testInvalidFilenameCountUnitThrowsException()
+    {
+        self::expectException(InvalidArgumentException::class);
+        self::expectExceptionMessage(\sprintf('The "filenameCountUnit" option must be one of the "%s::FILENAME_COUNT_*" constants ("%s" given).', File::class, 'nonExistentCountUnit'));
+        $file = new File(filenameCountUnit: 'nonExistentCountUnit');
     }
 
     /**
@@ -162,6 +188,9 @@ class FileTest extends TestCase
         self::assertSame(100000, $cConstraint->maxSize);
         self::assertSame(['my_group'], $cConstraint->groups);
         self::assertSame('some attached data', $cConstraint->payload);
+        self::assertSame(30, $cConstraint->filenameMaxLength);
+        self::assertSame('ISO-8859-15', $cConstraint->filenameCharset);
+        self::assertSame(File::FILENAME_COUNT_CODEPOINTS, $cConstraint->filenameCountUnit);
     }
 }
 
@@ -173,6 +202,6 @@ class FileDummy
     #[File(maxSize: 100, notFoundMessage: 'myMessage')]
     private $b;
 
-    #[File(maxSize: '100K', groups: ['my_group'], payload: 'some attached data')]
+    #[File(maxSize: '100K', filenameMaxLength: 30, filenameCharset: 'ISO-8859-15', filenameCountUnit: File::FILENAME_COUNT_CODEPOINTS, groups: ['my_group'], payload: 'some attached data')]
     private $c;
 }